Simplification
- abstract bulk string writing to functions
- remove ancillary functions to external files
- remove unused libraries
